feat: update opengraph meta tags (#88)
* update home page metadata


* add metadata for osnap page


* try absolute urls


* try different aspect ratio


* add twitter card images


* remove test image


---------
